+++ src/ResourceNotFoundException.php
@@ -11,6 +11,7 @@
 
 namespace Puli\Repository;
 
+use Exception;
 use RuntimeException;
 
 /**
@@ -21,4 +22,20 @@ use RuntimeException;
  */
 class ResourceNotFoundException extends RuntimeException
 {
+    /**
+     * Creates a new exception for a resource path.
+     *
+     * @param string    $path     The path which was not found.
+     * @param int       $code     The error code.
+     * @param Exception $previous The exception that caused this exception.
+     *
+     * @return static The created exception.
+     */
+    public static function forPath($path, $code = 0, Exception $previous = null)
+    {
+        return new static(sprintf(
+            'The resource %s does not exist.',
+            $path
+        ), $code, $previous);
+    }
 }